redis/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
azure-core==1.5.0
azure-eventhub==5.1.0
azure-eventhub-checkpointstoreblob-aio==1.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.13
multidict==4.7.5
oauthlib==3.1.0
pkg-resources==0.0.0
pycparser==2.20
redis==3.5.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
uamqp==1.2.7
urllib3==1.25.9
Werkzeug==1.0.1
yarl==1.4.2
